added steam Fixed clogged ui thread

+++ Minigin/Subject.cpp
@@ -3,7 +3,7 @@
 
 void Subject::NotifyAllObservers(const dae::GameObject& entity, eEvent event)
 {
-	for (int i = 0; i < m_pObservers.size(); ++i)
+	for (size_t i = 0; i < m_pObservers.size(); ++i)
 	{
 		m_pObservers[i]->OnNotify(entity, event);
 	}
